Refactor: change directory structure of domain
